CreateBoardModal　切り出し
