+++ src/components/index.ts
@@ -9,3 +9,4 @@ export { SnackbarProvider } from './common/SnackbarProvider'
 export { Board } from './views/Board'
 export { Home } from './views/Home'
 export { BoardTitle } from './board/BoardTitle'
+export { CreateBoardModal } from './board/CreateBoardModal'
